+++ src/packages/pulltorefresh/doc.zh-TW.md
@@ -16,28 +16,7 @@ import { PullToRefresh } from '@nutui/nutui-react';
 
 :::demo
 
-```tsx
-import React, {useState} from "react";
-import { PullToRefresh, Cell } from '@nutui/nutui-react'
-
-const App = () => {
-  const [list] = useState([1, 2, 3, 4, 5, 6, 7])
-  return (
-    <>
-      <div className="demo">
-        <h2>基礎用法</h2>
-        <PullToRefresh>
-          {list.map((item) => (
-            <Cell key={item}>{item}</Cell>
-          ))}
-        </PullToRefresh>
-      </div>
-    </>
-  )
-}
-
-export default App
-```
+<CodeBlock src='h5/demo1.tsx'></CodeBlock>
 
 :::
